LayoutTests/platform/mac-yosemite/fast/canvas/pattern-too-large-to-create-expected.txt

CONSOLE MESSAGE: line 18: Canvas area exceeds the maximum limit (width * height > 67108864).
PASS: Saw exception.
